Identify the difference between & and && in C.
1.Knowledge
Single ampersand & ===> Bit-wise and. 1000 & 1101 = 1010
Double ampersand && ===> Logical and. True & true = ture. None-zero value = true.
2.Two ways to load sample data:
Set a global/local 2-dimensional array for integers. Applied here.
Define a structure and redefine each time in the loop. Included in the comments but it's not recommended. 
3.Summary: 
On 20 Feb 2024 this 50-line c program was written as the first project in Github either as a remainder for myself when meeting similar problems again and an example of my coding. Not so much surprise, but happy enough for a freshman here.